ml-pipeline/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
scikit-learn==1.3.0
numpy==1.25.2
werkzeug==2.3.7
sentence-transformers==2.2.2
torch==2.0.1
